[FIX] #251 Nodes are now selected again after pasted
